+++ steam.trade.offer.enhancer.user.js
@@ -1,7 +1,7 @@
 // ==UserScript==
 // @name        Steam Trade Offer Enhancer
 // @description Browser script to enhance Steam trade offers.
-// @version     2.1.1
+// @version     2.1.2
 // @author      Julia
 // @namespace   http://steamcommunity.com/profiles/76561198080179568/
 // @updateURL   https://github.com/juliarose/steam-trade-offer-enhancer/raw/master/steam.trade.offer.enhancer.meta.js
@@ -1271,6 +1271,10 @@
                     color: #FFFFFF;
                 }
                 
+                #add_btns .btn_small > span {
+                    padding: 0 12px;
+                }
+                
                 .btn_green {
                     background-color: #709D3C;
                 }
@@ -1515,7 +1519,7 @@
                                 const imgs = [`url(${img})`];
                                 const classes = ['summary_item'];
                                 
-                                if (effect !== 'none') {
+                                if (effect !== undefined && effect !== 'none') {
                                     imgs.push(`url('${getEffectURL(effect)}')`);
                                 }
                                 
@@ -1988,18 +1992,9 @@
                      * @returns {Array} Array of picked items.
                      */
                     function pickItems(you, amount, index, finder) {
-                        // get inventory for selected app and context
-                        function getInventory(user) {
-                            return (user.rgAppInfo[appid] &&
-                                user.rgAppInfo[appid].rgContexts[contextid].inventory &&
-                                user.rgAppInfo[appid].rgContexts[contextid].inventory.rgInventory
-                            ) || {};
-                        }
-                        
                         function getItems(you) {
-                            const user = you ? UserYou : UserThem;
                             const $items = (you ? page.$yourSlots : page.$theirSlots).find('.item');
-                            const inventory = getInventory(user);
+                            const inventory = getInventory(appid, contextid, you);
                             // get ids of items in trade offer matching app
                             const addedIDs = $items.toArray().reduce((arr, el) => {
                                 const split = el.id.replace('item', '').split('_');
@@ -2056,9 +2051,7 @@
                             return items;
                         }
                         
-                        const $inventory = page.get.$inventory();
-                        const match = ($inventory.attr('id') || '').match(/(\d+)_(\d+)$/);
-                        const [ , appid, contextid] = (match || []);
+                        const { appid, contextid } = getInventoryApp();
                         
                         // inventory must be present
                         if (!appid) {
@@ -2219,6 +2212,75 @@
                                     satisfied
                                 };
                             },
+                            // get items by whether they were recently obtained
+                            'RECENT': function() {
+                                // gets nearest numbers to a given number within range of gap
+                                const getNearNumbers = (nums, near, gap) => {
+                                    if (nums.length === 0) {
+                                        return [];
+                                    }
+                                    
+                                    const getDistance = (num) => Math.abs(num - near);
+                                    const sorted = nums
+                                        // add distance from "near" for each num
+                                        .map((num) => {
+                                            return {
+                                                num,
+                                                distance: getDistance(num)
+                                            };
+                                        })
+                                        .sort((a, b) => {
+                                            return a.distance - b.distance;
+                                        });
+                                    
+                                    // check if the nearest value is within the gap value
+                                    if (sorted[0].distance > gap) {
+                                        return [];
+                                    }
+                                    
+                                    // add the initial
+                                    const values = [sorted[0].num];
+                                    
+                                    // loop through sorted values
+                                    for (let i = 1; i < sorted.length; i++) {
+                                        const current = sorted[i];
+                                        const prev = sorted[i - 1];
+                                        const difference = Math.abs(prev.distance - current.distance);
+                                        
+                                        // gap is too big
+                                        if (difference > gap) {
+                                            // stop bleeding
+                                            return values;
+                                        }
+                                        
+                                        values.push(current.num);
+                                    }
+                                    
+                                    return values;
+                                };
+                                const { appid, contextid } = getInventoryApp();
+                                const inventories = (
+                                    you === null ?
+                                        [true, false] :
+                                        [you]
+                                ).map(you => getInventory(appid, contextid, you));
+                                const ids = inventories.reduce((ids, inventory) => {
+                                    return [
+                                        ...ids,
+                                        ...Object.keys(inventory).map(id => parseInt(id))
+                                    ];
+                                }, []);
+                                const highestId = Math.max(0, ...ids);
+                                const nearIds = getNearNumbers(ids, highestId, 50).map(id => id.toString());
+                                const found = finders.id(nearIds);
+                                const items = getElementsForItems(found);
+                                const satisfied = nearIds.length === items.length;
+                                
+                                return {
+                                    items,
+                                    satisfied
+                                };
+                            },
                             // get items displayed in the inventory
                             'ITEMS': function() {
                                 // check if an items is visible on page
@@ -2250,6 +2312,29 @@
                     
                     return getItems;
                 }());
+                
+                // get inventory for selected app and context of user
+                function getInventory(appid, contextid, you) {
+                    const user = you ? UserYou : UserThem;
+                    
+                    return (
+                        user.rgAppInfo[appid] &&
+                        user.rgAppInfo[appid].rgContexts[contextid].inventory &&
+                        user.rgAppInfo[appid].rgContexts[contextid].inventory.rgInventory
+                    ) || {};
+                }
+                
+                // gets the app of the currently visible inventory
+                function getInventoryApp() {
+                    const $inventory = page.get.$inventory();
+                    const match = ($inventory.attr('id') || '').match(/(\d+)_(\d+)$/);
+                    const [ , appid, contextid] = (match || []);
+                    
+                    return {
+                        appid,
+                        contextid
+                    };
+                }
                 
                 // customizes the elements within this inventory
                 function customizeItems(inventory) {
@@ -2287,6 +2372,9 @@
                                 </div>
                                 <div id="btn_addmetal" class="btn_silver btn_black btn_small">
                                     <span>Add Metal</span>
+                                </div>
+                                <div id="btn_addrecent" class="btn_silver btn_black btn_small">
+                                    <span>Add Recent</span>
                                 </div>
                                 <div id="btn_addlisting" class="btn_blue btn_black btn_small">
                                     <span>Add Listing</span>
@@ -2358,6 +2446,7 @@
                         $items: $('#btn_additems'),
                         $keys: $('#btn_addkeys'),
                         $metal: $('#btn_addmetal'),
+                        $recent: $('#btn_addrecent'),
                         $listing: $('#btn_addlisting'),
                         $addIDs: $('#btn_addids'),
                         $getIDs: $('#btn_getids')
@@ -2550,6 +2639,9 @@
                     });
                     page.btns.$metal.on('click', () => {
                         addItems('METAL', ...getDefaults());
+                    });
+                    page.btns.$recent.on('click', () => {
+                        addItems('RECENT', ...getDefaults());
                     });
                     page.btns.$listing.on('click', () => {
                         addListingPrice();
@@ -2902,7 +2994,7 @@
     (function() {
         const DEPS = (function() {
             // current version number of script
-            const VERSION = '2.1.1';
+            const VERSION = '2.1.2';
             // our window object for accessing globals
             const WINDOW = unsafeWindow;
             // dependencies to provide to each page script